fix: JS pointer-drag scroll for popup content areas and selector grids
Extracts the tab-strip drag logic into a shared `scrollDrag.ts` utility
and wires it into every dialog opened by `dialogManager.ts`, targeting:

- `.PopupContent` (vertical axis) — tile grids, token lists, karma/upgrade
  choosers; CSS `touch-action: pan-x` frees the vertical axis for the JS
  handler, mirroring how the tab strip uses `touch-action: pan-y` for
  horizontal drag
- `.PopupSelector` and `.PopupTokens` (horizontal axis) — paged grids;
  `touch-action: pan-y` mirrors the existing tab-strip approach exactly

PopupMenu.tsx is refactored to use the shared utility (no behaviour change).

Reverts the button overflow change from the previous commit (the
`position: static; flex-wrap` layout on `.PopupButtons`) and restores the
`padding-bottom: 76px` straddle geometry; button layout will be addressed
separately after agreeing on the right visual approach.
